chatbotAI/requirements.txt
flask==3.0.2
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.31
langchain-google-genai==0.0.11
langchain-pinecone==0.2.3
pinecone-client==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
gunicorn==21.2.0